+++ site/theme/mdx/ComponentIntro.tsx
@@ -1,1 +1,0 @@
-export { ComponentIntro as default } from "../components/ComponentIntro";